auto-researcher/pyproject.toml
[project]
name = "auto-researcher-tools"
version = "0.1.0"
description = "Validation helpers for the auto-researcher Cursor skill (PROGRAM.md parsing and metric extraction)."
readme = "SKILL.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts"]
addopts = "-q"
